Add/delete ingredient (en attendant d'ajouter le formulaire)

+++ www/src/lbs/api/gestionController.php
@@ -20,37 +20,49 @@ Class gestionController extends baseController
 	}
 
 	//supprimer un ingrédient dans la liste
-	public function suppIngredient(Request $request, Response $response, $args){
-		try{
-			$ingredient = Ingredient::select()->where('id','=',$args['id'])->firstOrFail();
-			$ingredient->delete();
-			$response = $this->json_success($response, 201, "deletion done");
-		}catch(ModelNotFoundException $e){
-			$response = $response>withStatus(404)->withHeader('Content-type', 'application/json');
-			$errorMessage = ["error" => "id not found" ];
-			$response->getBody()->write(json_encode($errorMessage));
+	public function suppIngredient($request, $response, $args){
+		try {
+
+			$ingredient = Ingredient::select()->where('id','=',$args['id'])->first();
+			if (!empty($ingredient)) {
+				$ingredient->delete();
+				$response = $response->withJson(array('Succes' => "deleted"), 201);
+			} 
+
+			$response = $response->withJson(array('Erreur' => "Not found"), 404);
+
+		} catch(ModelNotFoundException $e){
+
+			$response = $response->withJson(array('Erreur' => "Not found"), 404);
+
 		}
+
 		return $response;
 	}
 
-
 	//ajouter un ingredient
-	public function ajouterIngredient(Request $request, Response $response, $args){
+	public function ajouterIngredient($request, $response, $args)
+	{
 		$ingredient = new Ingredient;
-		if(isset($request->getParsedBody()['cat_id'])){
-				 $ingredient->nom = filter_var($request->getParsedBody()['nom'], FILTER_SANITIZE_STRING);
-				 $ingredient->cat_id = filter_var($request->getParsedBody()['cat_id'], FILTER_SANITIZE_INT);
-				 $ingredient->description = filter_var($request->getParsedBody()['description'], FILTER_SANITIZE_STRING);
-				 $ingredient->fournisseur = filter_var($request->getParsedBody()['fournisseur'], FILTER_SANITIZE_STRING);
-				 $ingredient->img = filter_var($request->getParsedBody()['img'], FILTER_UNSAFE_RAW);
-				 $ingredient->save();
+		$body = $request->getParsedBody();
 
-				 $newIngredient=array($ingredient, $ingredient->getingredientcat());
+		if(isset($body['cat_id']))
+		{
+			$ingredient->nom = filter_var($body['nom'], FILTER_SANITIZE_STRING);
+			$ingredient->cat_id = filter_var($body['cat_id'], FILTER_SANITIZE_NUMBER_INT);
+			$ingredient->description = filter_var($body['description'], FILTER_SANITIZE_STRING);
+			$ingredient->fournisseur = filter_var($body['fournisseur'], FILTER_SANITIZE_STRING);
+			$ingredient->img = filter_var($body['img'], FILTER_UNSAFE_RAW);
+			$ingredient->save();
 
-				 $response = $this->json_success($response, 201, $newIngredient->toJson());
-		}else{
-			 $response = $this->json_error($response, 500, "erreur lors de la creation de la ressource");
+			$newIngredient = array($ingredient, $ingredient->categorie);
+
+			$response = $response->withJson($newIngredient->toJson(), 201);
+
+		} else {
+			$response = $response->withJson(array('Erreur' => "erreur lors de la creation de la ressource"), 500);
 		}
+
 		return $response;
 	}
